+++ src/main/java/kitchenpos/dto/MenuProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.dto;
 
 public class MenuProductResponse {
     private Long seq;
