+++ src/main/java/kitchenpos/dto/OrderLineItemRequest.java
@@ -1,9 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.dto;
 
-public class MenuProductRequest {
+public class OrderLineItemRequest {
     private Long seq;
+    private Long orderId;
     private Long menuId;
-    private Long productId;
     private long quantity;
 
     public Long getSeq() {
@@ -14,20 +14,20 @@ public class MenuProductRequest {
         this.seq = seq;
     }
 
+    public Long getOrderId() {
+        return orderId;
+    }
+
+    public void setOrderId(final Long orderId) {
+        this.orderId = orderId;
+    }
+
     public Long getMenuId() {
         return menuId;
     }
 
     public void setMenuId(final Long menuId) {
         this.menuId = menuId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public void setProductId(final Long productId) {
-        this.productId = productId;
     }
 
     public long getQuantity() {
