+++ src/main/java/kitchenpos/dto/OrderRequest.java
@@ -1,9 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.dto;
+
+import kitchenpos.domain.OrderLineItem;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class OrderResponse {
+public class OrderRequest {
     private Long id;
     private Long orderTableId;
     private String orderStatus;
